Cortina/02. Service/Draft 1/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
pillow
numpy
scipy
tensorflow
keras
